requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic<2.0.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
langchain==0.1.20
langchain-community==0.0.38
langchain-openai
azure-cosmos==4.5.1
openai==1.12.0
python-multipart==0.0.6